+++ src/main/java/com/verdantartifice/primalmagick/client/compat/jei/dissolution/DissolutionRecipeCategory.java
@@ -1,0 +1,84 @@
+package com.verdantartifice.primalmagick.client.compat.jei.dissolution;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.verdantartifice.primalmagick.PrimalMagick;
+import com.verdantartifice.primalmagick.client.compat.jei.RecipeCategoryPM;
+import com.verdantartifice.primalmagick.common.crafting.IDissolutionRecipe;
+import com.verdantartifice.primalmagick.common.items.ItemsPM;
+import com.verdantartifice.primalmagick.common.sources.Source;
+import com.verdantartifice.primalmagick.common.sources.SourceList;
+
+import mezz.jei.api.constants.VanillaTypes;
+import mezz.jei.api.gui.IRecipeLayout;
+import mezz.jei.api.gui.drawable.IDrawableStatic;
+import mezz.jei.api.gui.ingredient.IGuiItemStackGroup;
+import mezz.jei.api.helpers.IGuiHelper;
+import mezz.jei.api.ingredients.IIngredients;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.ItemStack;
+
+/**
+ * Recipe category for a dissolution recipe.
+ * 
+ * @author Daedalus4096
+ */
+public class DissolutionRecipeCategory extends RecipeCategoryPM<IDissolutionRecipe> {
+    public static final ResourceLocation UID = new ResourceLocation(PrimalMagick.MODID, "dissolution_chamber");
+    private static final ResourceLocation BACKGROUND_TEXTURE = new ResourceLocation(PrimalMagick.MODID, "textures/gui/jei/dissolution_chamber.png");
+    private static final int MANA_COST_X_OFFSET = 28;
+    private static final int MANA_COST_Y_OFFSET = 1;
+
+    private final IDrawableStatic manaCostIcon;
+    
+    public DissolutionRecipeCategory(IGuiHelper guiHelper) {
+        super(IDissolutionRecipe.class, guiHelper, UID, "block.primalmagick.dissolution_chamber");
+        this.manaCostIcon = guiHelper.createDrawable(BACKGROUND_TEXTURE, 82, 0, 16, 16);
+        this.setBackground(guiHelper.createDrawable(BACKGROUND_TEXTURE, 0, 0, 82, 44));
+        this.setIcon(new ItemStack(ItemsPM.DISSOLUTION_CHAMBER.get()));
+    }
+
+    @Override
+    public void setIngredients(IDissolutionRecipe recipe, IIngredients ingredients) {
+        ingredients.setInputIngredients(recipe.getIngredients());
+        ingredients.setOutput(VanillaTypes.ITEM, recipe.getResultItem());
+    }
+
+    @Override
+    public void setRecipe(IRecipeLayout recipeLayout, IDissolutionRecipe recipe, IIngredients ingredients) {
+        IGuiItemStackGroup guiItemStacks = recipeLayout.getItemStacks();
+
+        guiItemStacks.init(0, true, 0, 18);
+        guiItemStacks.init(1, false, 60, 18);
+
+        guiItemStacks.set(ingredients);
+    }
+
+    @Override
+    public void draw(IDissolutionRecipe recipe, PoseStack stack, double mouseX, double mouseY) {
+        if (recipe.getManaCosts() != null && !recipe.getManaCosts().isEmpty()) {
+            this.manaCostIcon.draw(stack, MANA_COST_X_OFFSET, MANA_COST_Y_OFFSET);
+        }
+    }
+
+    @Override
+    public List<Component> getTooltipStrings(IDissolutionRecipe recipe, double mouseX, double mouseY) {
+        SourceList manaCosts = recipe.getManaCosts();
+        if ( manaCosts != null && !manaCosts.isEmpty() && 
+             mouseX >= MANA_COST_X_OFFSET && mouseX < MANA_COST_X_OFFSET + this.manaCostIcon.getWidth() &&
+             mouseY >= MANA_COST_Y_OFFSET && mouseY < MANA_COST_Y_OFFSET + this.manaCostIcon.getHeight() ) {
+            List<Component> tooltip = new ArrayList<>();
+            tooltip.add(new TranslatableComponent("primalmagick.crafting.mana_cost_header"));
+            for (Source source : manaCosts.getSourcesSorted()) {
+                tooltip.add(new TranslatableComponent("primalmagick.crafting.mana_tooltip", manaCosts.getAmount(source), source.getNameText()));
+            }
+            return tooltip;
+        } else {
+            return super.getTooltipStrings(recipe, mouseX, mouseY);
+        }
+    }
+}