+++ changelog.md
@@ -1,6 +1,7 @@
 v2.0.3
 ------
 * Update model for runecarving table
+* Fix misaligned mana cost widgets on Arcane Workbench screen
 
 v2.0.2
 ------